+++ angular-app/ngx-fittable-docs/src/app/topics/table-designer/custom-view/custom-view.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, ViewEncapsulation } from '@angular/core';
 
 import { createTable, registerModelConfig } from 'fittable-core/model';
 import { registerOperationConfig } from 'fittable-core/operations';
@@ -21,7 +21,12 @@ import { TopicTitle } from '../../../common/topic-title.model';
 @Component({
   selector: 'custom-view',
   templateUrl: './custom-view.component.html',
-  styleUrls: ['./custom-view.component.css', '../../common/common.css'],
+  styleUrls: [
+    './custom-view.component.css',
+    '../../common/common.css',
+    '../../../../../node_modules/fittable-angular/src/lib/common/css/fittable-main.css',
+  ],
+  encapsulation: ViewEncapsulation.None,
 })
 export class CustomViewComponent {
   public title: TopicTitle = 'Custom view';