+++ angular-lib/projects/fittable-angular/src/public-api.ts
@@ -4,8 +4,8 @@
 
 export * from './lib/fittable.module';
 export * from './lib/fittable.component';
-export * from './lib/controls-view/toolbar/toolbar.component';
 export * from './lib/table-view/table.component';
+export * from './lib/controls-view/toolbar/toolbar.component';
 export * from './lib/controls-view/statusbar/statusbar.component';
 export * from './lib/controls-view/settings-bar/settings-bar.component';
 export * from './lib/controls-view/context-menu/context-menu.component';